Refactor: extract tryParse
